Feature/bug the application gets the same data for all users #47

+++ controllers/todo.js
@@ -10,7 +10,7 @@ const createTodo = async(req,res)=>{
 
 const updateTodo= async (req,res)=>{
     const { todoId } = req.params;
-    const todo = await Todo.findOneAndUpdate({_id:todoId}, req.body,{new:true})
+    const todo = await Todo.findOneAndUpdate({_id:todoId,createdBy:req.userId}, req.body,{new:true})
     if(!todo){
         throw new NotFoundError(`Todo Item with: ${todoId} does found`)
     }
@@ -19,14 +19,14 @@ const updateTodo= async (req,res)=>{
 
 
 const getAllTodo = async (req,res)=>{
-    const todoList = await Todo.find({});
+    const todoList = await Todo.find({createdBy:req.userId});
     res.status(200).json({data:todoList});
 }
 
 const getTodo = async (req,res)=>{
     //const {userId} = req.user;
     const { todoId }= req.params;
-    const todo = await Todo.findOne({_id:todoId});
+    const todo = await Todo.findOne({_id:todoId,createdBy:req.userId});
     if(!todo){
         throw new NotFoundError(`Todo with id: ${todoId}, Not Found`)
     }
@@ -36,7 +36,7 @@ const getTodo = async (req,res)=>{
 const deleteTodo = async (req,res)=>{
     //const { userId} = req.user;
     const { todoId} = req.params;
-    const todo = await Todo.deleteOne({_id:todoId})
+    const todo = await Todo.deleteOne({_id:todoId,createdBy:req.userId})
     if(!todo){
         throw new NotFoundError(`Todo with id: ${todoId}, Not Found`)
     }
